+++ Samples~/Scripts/IssueAndTesting/Issue/Issue8/Anim.cs
@@ -5,6 +5,7 @@ namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue8
     public class Anim : AnimBase
     {
         [AnimatorParam(nameof(Animator), AnimatorControllerParameterType.Bool)] public int animParam;
+        [AnimatorParam(nameof(Animator))] public int animParam2;
         [AnimatorState(nameof(Animator))] public string anim;
     }
 }